requirements.txt
PySide6>=6.7
pyserial>=3.5
opencv-python>=4.9
numpy>=1.24
scipy>=1.11
matlabengine==25.2
